Use stride_iterator instead of converting to view1
We need to get the array index of the current element for functions
like for_each_vertex. Was using a hack to do this: convert the
element pointer back to view-2 and subtracting the begin() pointer
of the array. Turns out calling view2to1 for each iteration of the
worker function is really expensive. So we're back to using
stride_iterator, which allows simple subtraction.